:four_leaf_clover: category info.

+++ src/main/java/cn/zealon/notes/service/CategoryService.java
@@ -5,12 +5,11 @@ import cn.zealon.notes.common.result.ResultUtil;
 import cn.zealon.notes.common.utils.DateUtil;
 import cn.zealon.notes.domain.Category;
 import cn.zealon.notes.repository.CategoryRepository;
+import cn.zealon.notes.repository.NotesRepository;
 import cn.zealon.notes.security.jwt.JwtAuthService;
-import cn.zealon.notes.vo.CategoryCascadeVO;
 import cn.zealon.notes.vo.CategoryVO;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
-import org.bson.types.ObjectId;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Sort;
@@ -36,13 +35,11 @@ public class CategoryService {
     @Autowired
     private CategoryRepository categoryRepository;
 
+    @Autowired
+    private NotesRepository notesRepository;
+
     public Result create(Category category) {
         try {
-            int level = 1;
-            if (StringUtils.isNotBlank(category.getParentId())) {
-                level = 2;
-            }
-            category.setLevel(level);
             category.setUserId(jwtAuthService.getLoginUserBean().getUsername());
             Category insert = this.categoryRepository.insert(category);
             return ResultUtil.success(insert);
@@ -54,18 +51,6 @@ public class CategoryService {
 
     public Result update(Category category) {
         try {
-            // 业务校验
-            if (StringUtils.isNotBlank(category.getParentId())) {
-                Category parentCategory = this.categoryRepository.findOne(category.getParentId());
-                if (parentCategory != null && parentCategory.getLevel() == 2) {
-                    return ResultUtil.verificationFailed().buildMessage("分类深度最多是两层哦！");
-                }
-
-                if (this.categoryRepository.getSubCategoryCountById(category.getId()) > 0) {
-                    return ResultUtil.verificationFailed().buildMessage("当前子分类有数据哦，请先移动子分类到其它分类下再调整父分类！");
-                }
-            }
-
             String nowDateString = DateUtil.getNowDateString();
             Update update = Update.update("update_time", nowDateString);
             if (StringUtils.isNotBlank(category.getTitle())) {
@@ -74,12 +59,6 @@ public class CategoryService {
             if (category.getSort() != null) {
                 update.set("sort", category.getSort());
             }
-            int level = 1;
-            if (StringUtils.isNotBlank(category.getParentId())) {
-                level = 2;
-            }
-            update.set("level", level);
-            update.set("parent_id", category.getParentId());
             this.categoryRepository.updateOne(category.getId(), update);
             return ResultUtil.success();
         } catch (Exception ex) {
@@ -90,6 +69,10 @@ public class CategoryService {
 
     public Result remove(Category category){
         try {
+            long notesCount = this.notesRepository.findCountByCategoryId(category.getId());
+            if (notesCount > 0) {
+                return ResultUtil.verificationFailed().buildMessage("分类下有"+notesCount+"个笔记，请移动笔记到其它分类下再删除！");
+            }
             this.categoryRepository.remove(category.getId());
             return ResultUtil.success();
         } catch (Exception ex) {
@@ -99,26 +82,16 @@ public class CategoryService {
     }
 
     /**
-     * 获取用户全部分类(深度2)
+     * 获取用户全部分类
      * @return
      */
     public Result getAllCategoryList() {
-        String parentId = "";
         String userId = jwtAuthService.getLoginUserBean().getUsername();
-        List<Category> categories = this.getCategoryListByParentId(userId, parentId);
+        List<Category> categories = this.getCategoryListByParentId(userId);
         List<CategoryVO> categoryVOS = new ArrayList<>();
         for (Category category : categories) {
             CategoryVO vo = new CategoryVO();
             BeanUtils.copyProperties(category, vo);
-            List<Category> subCategories = this.getCategoryListByParentId(userId, category.getId());
-            List<CategoryVO> subCategoryVOS = new ArrayList<>();
-            for (Category sub : subCategories){
-                CategoryVO subVo = new CategoryVO();
-                BeanUtils.copyProperties(sub, subVo);
-                subVo.setCategorys(new ArrayList<>());
-                subCategoryVOS.add(subVo);
-            }
-            vo.setCategorys(subCategoryVOS);
             categoryVOS.add(vo);
         }
         return ResultUtil.success(categoryVOS);
@@ -129,26 +102,15 @@ public class CategoryService {
      * @param id
      * @return
      */
-    public Result getCascadeCategoryNames(String id) {
-        CategoryCascadeVO vo = new CategoryCascadeVO();
+    public Result getCategoryDetails(String id) {
+        CategoryVO vo = new CategoryVO();
         Category category = this.categoryRepository.findOne(id);
-        if (category != null) {
-            Category category2 = this.categoryRepository.findOne(category.getParentId());
-            if (category2 != null) {
-                vo.setParentName(category2.getTitle());
-            }
-            vo.setName(category.getTitle());
-        }
+        BeanUtils.copyProperties(category, vo);
         return ResultUtil.success(vo);
     }
 
-    private List<Category> getCategoryListByParentId(String userId, String parentId) {
+    private List<Category> getCategoryListByParentId(String userId) {
         Query query = Query.query(Criteria.where("user_id").is(userId));
-        if (StringUtils.isBlank(parentId)) {
-            parentId = "";
-
-        }
-        query.addCriteria(Criteria.where("parent_id").is(parentId));
         query.with(Sort.by("sort").ascending());
         return this.categoryRepository.find(query);
     }
